requirements.txt
azure-core==1.8.1
azure-cosmos==4.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
MarkupSafe==1.1.1
nltk==3.5
regex==2020.9.27
requests==2.24.0
six==1.15.0
tqdm==4.50.0
urllib3==1.25.10
Werkzeug==1.0.1
